pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civic-volunteering-monitor"
version = "0.1.0"
description = "Deterministic monitor for municipal volunteering policy documents"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "pandas>=2.1",
  "openpyxl>=3.1",
  "pypdf>=4.2",
  "python-dotenv>=1.0",
  "openai>=1.35"
]

[project.optional-dependencies]
docx = ["python-docx>=1.1"]
playwright = ["playwright>=1.45"]

[project.scripts]
monitor = "monitor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
